Moved UsersPaged from patient package to user package. New ModelBase class with UUID, creation and update time. Extend Patient and User from ModelBase class. Remove UUID from Patient and User because is in ModelBase
