+++ src/main/java/com/systelab/seed/model/patient/Patient.java
@@ -1,12 +1,13 @@
 package com.systelab.seed.model.patient;
 
+import com.systelab.seed.model.ModelBase;
 import com.systelab.seed.util.constraints.Email;
 import com.systelab.seed.util.convert.jaxb.JsonLocalDateTypeAdapter;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.*;
 import javax.validation.constraints.Size;
@@ -15,27 +16,21 @@ import javax.xml.bind.annotation.XmlType;
 import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 import java.io.Serializable;
 import java.time.LocalDate;
-import java.util.UUID;
 
 @XmlRootElement
-@XmlType(propOrder = {"id", "name", "surname", "email", "dob", "address"})
+@XmlType(propOrder = {"id", "creationTime", "updateTime", "name", "surname", "medicalNumber", "email", "dob", "address"})
 
 @Entity
 @Data
+@EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
 @AllArgsConstructor
 @Table(name = "patient")
 @NamedQueries({@NamedQuery(name = Patient.FIND_ALL, query = "SELECT p FROM Patient p"),
         @NamedQuery(name = Patient.ALL_COUNT, query = "SELECT COUNT(p.id) FROM Patient p")})
-public class Patient implements Serializable {
+public class Patient extends ModelBase implements Serializable {
     public static final String FIND_ALL = "Patient.findAll";
     public static final String ALL_COUNT = "Patient.allCount";
-
-    @Id
-    @GeneratedValue(generator = "uuid2")
-    @GenericGenerator(name = "uuid2", strategy = "uuid2")
-    @Column(columnDefinition = "BINARY(16)")
-    private UUID id;
 
     @Size(min = 1, max = 255)
     private String surname;
