+++ src/main/java/com/systelab/seed/model/user/UsersPage.java
@@ -1,4 +1,4 @@
-package com.systelab.seed.model.patient;
+package com.systelab.seed.model.user;
 
 import com.systelab.seed.model.user.User;
 import com.systelab.seed.util.pagination.Page;
